+++ examples/010-editing-gitignore/.gitignore
@@ -1,0 +1,5 @@
+.pytest_cache/
+__pycache__/
+*.egg-info/
+/dist/
+dist/
