+++ examples/010-editing-gitignore/.project-config.toml
@@ -1,0 +1,1 @@
+style = "style.json5"
