+++ examples/010-editing-gitignore/README.rst
@@ -1,0 +1,5 @@
+..
+   Name: Editing a .gitignore file
+   Exitcode: 0
+
+Enforce the existence of certain lines in a `.gitignore` file.
