+++ examples/011-replace-codeblocks-langs/.project-config.toml
@@ -1,0 +1,1 @@
+style = "style.json5"
