+++ examples/011-replace-codeblocks-langs/README.rst
@@ -1,0 +1,8 @@
+..
+   Name: Replacing code blocks languages in RST documents
+   Exitcode: 0
+
+Don't allow code blocks in RST documentation files:
+
+* Bash is not a POSIX compliant shell, use Shell lexer.
+* Pygments' JSON5 lexer is not implemented yet, use Javascript lexer.
