+++ examples/011-replace-codeblocks-langs/file.rst
@@ -1,0 +1,12 @@
+.. code-block::  python
+
+   foo = "bar"
+
+.. code-block:: json5
+
+   {}
+
+.. code-block:: sh
+
+   #!/bin/sh
+   curl -X POST http://localhost:8080/api/v1/users
