+++ examples/_001-basic-usage-fails-file-content/README.rst
@@ -1,5 +1,4 @@
 ..
    Name: Basic usage
    Exitcode: 1
-   Stderr: .gitignore\n  - Expected line '/build/' not found rules[0].includeLines[0]
-   Fixable: false
+   Stderr: .gitignore\n  - (FIXABLE) Expected line '/build/' not found rules[0].includeLines[0]
