+++ examples/_002-assert-project-config-self-config-missing-cache/README.rst
@@ -1,8 +1,7 @@
 ..
    Name: project-config self configuration
    Exitcode: 1
-   Stderr: .project-config.toml\n  - JMESPath 'contains(keys(@), 'cache')' does not match. Expected True, returned False rules[0].JMESPathsMatch[0] Configure the cache for project-config explicitly
-   Fixable: true
+   Stderr: .project-config.toml\n  - (FIXABLE) JMESPath 'contains(keys(@), 'cache')' does not match. Expected True, returned False rules[0].JMESPathsMatch[0] Configure the cache for project-config explicitly
 
 The ``cache`` configuration field is being explicitly defined in
 the **project-config** configuration.
