+++ examples/_007-jmespath-match-against-online-source-fails/README.rst
@@ -3,6 +3,7 @@
    Exitcode: 1
    Stderr: package.json\n  - JMESPath 'contains([1], [0])' does not match. Expected True, returned False rules[0].crossJMESPathsMatch[0]
    Online: true
+   Fixable: false
 
 Check that the ``license`` field of *package.json* file is defined with
 a valid OSI approved SPDX license identifier.
