+++ pyproject.toml
@@ -88,28 +88,28 @@ targets = [
 [tool.project-config]
 style = [
   # Basic Python styles: use poetry
-  "../project-config-styles/python/base.json5",
+  "gh://mondeja/project-config-styles/python/base.json5",
   # Minimum Python version
-  "../project-config-styles/python/version/min-37.json5",
+  "gh://mondeja/project-config-styles/python/version/min-37.json5",
   # Double quotes: formatting with Black
-  "../project-config-styles/python/double-quotes.json5",
+  "gh://mondeja/project-config-styles/python/double-quotes.json5",
   # Maximum line length
-  "../project-config-styles/python/line-length-80.json5",
+  "gh://mondeja/project-config-styles/python/line-length-80.json5",
   # Testing styles: use pytest
-  "../project-config-styles/python/tests.json5",
+  "gh://mondeja/project-config-styles/python/tests.json5",
   # Documentation styles: use Sphinx with Readthedocs theme
-  "../project-config-styles/python/sphinx.json5",
-  "../project-config-styles/python/google-docstrings.json5",
-  "../project-config-styles/python/readthedocs.json5",
+  "gh://mondeja/project-config-styles/python/sphinx.json5",
+  "gh://mondeja/project-config-styles/python/google-docstrings.json5",
+  "gh://mondeja/project-config-styles/python/readthedocs.json5",
   # Type checking with mypy
-  "../project-config-styles/python/mypy.json5",
+  "gh://mondeja/project-config-styles/python/mypy.json5",
 ]
 cache = "2 days"
 
 [tool.poetry-exec-plugin.commands]
 t = "coverage run -m pytest"
 test = "coverage run -m pytest"
-"test:show" = """coverage run -m pytest &&
+"test:show" = """pysetenv PROJECT_CONFIG_TESTS_E2E=true coverage run -m pytest &&
 coverage html &&
 python -c 'import webbrowser as w;w.open("http://127.0.0.1:8088")' &&
 python -m http.server 8088 -b localhost -d htmlcov
