+++ src/project_config/plugins/inclusion.py
@@ -15,6 +15,11 @@ from project_config import (
     Rule,
     Tree,
 )
+from project_config.utils.jmespath import (
+    JMESPathError,
+    compile_JMESPath_expression_or_error,
+    fix_tree_serialized_file_by_jmespath,
+)
 
 
 def _directories_not_accepted_as_inputs_error(
@@ -41,8 +46,6 @@ class InclusionPlugin:
         rule: Rule,
         context: ActionsContext,
     ) -> Results:
-        # TODO: allow to fix this rule serializing files as text
-        #   (array of lines)
         if not isinstance(value, list):
             yield InterruptingError, {
                 "message": "The value must be of type array",
@@ -58,11 +61,41 @@ class InclusionPlugin:
 
         expected_lines = []
         for i, line in enumerate(value):
-            if not isinstance(line, str):
+            fixer_query = ""
+
+            if isinstance(line, list):
+                # Fixer query expression
+                if len(line) != 2:
+                    yield InterruptingError, {
+                        "message": (
+                            "The '[expected-line, fixer_query]' array"
+                            f" '{pprint.pformat(line)}'"
+                            " must be of length 2"
+                        ),
+                        "definition": f".includeLines[{i}]",
+                    }
+                    return
+
+                line, fixer_query = line
+
+                if not isinstance(line, str) or not isinstance(
+                    fixer_query, str
+                ):
+                    yield InterruptingError, {
+                        "message": (
+                            "The '[expected-line, fixer_query]' array items"
+                            f" '{pprint.pformat([line, fixer_query])}'"
+                            " must be of type string"
+                        ),
+                        "definition": f".includeLines[{i}]",
+                    }
+                    return
+
+            elif not isinstance(line, str):
                 yield InterruptingError, {
                     "message": (
                         f"The expected line '{pprint.pformat(line)}'"
-                        " must be of type string"
+                        " must be of type string or array"
                     ),
                     "definition": f".includeLines[{i}]",
                 }
@@ -100,10 +133,51 @@ class InclusionPlugin:
             fcontent_lines = fcontent.splitlines()
             for line_index, expected_line in enumerate(expected_lines):
                 if expected_line not in fcontent_lines:
+                    if context.fix:
+                        if not fixer_query:
+                            fixer_query = f"insert(@, `-1`, '{expected_line}')"
+
+                        try:
+                            compiled_fixer_query = (
+                                compile_JMESPath_expression_or_error(
+                                    fixer_query,
+                                )
+                            )
+                        except JMESPathError as exc:
+                            yield InterruptingError, {
+                                "message": exc.message,
+                                "definition": f".includeLines[{line_index}]",
+                            }
+                            continue
+
+                        _, instance = tree.serialize_file(fpath)
+
+                        try:
+                            diff = fix_tree_serialized_file_by_jmespath(
+                                compiled_fixer_query,
+                                instance,
+                                fpath,
+                                tree,
+                            )
+                        except JMESPathError as exc:
+                            yield InterruptingError, {
+                                "message": exc.message,
+                                "definition": f".includeLines[{line_index}]",
+                            }
+                            continue
+                        else:
+                            fixed = True
+                            if not diff:
+                                continue
+                    else:
+                        fixed = False
+
                     yield Error, {
                         "message": f"Expected line '{expected_line}' not found",
                         "file": fpath,
                         "definition": f".includeLines[{line_index}]",
+                        "fixed": fixed,
+                        "fixable": True,
                     }
 
     @staticmethod
@@ -249,18 +323,36 @@ class InclusionPlugin:
             # Normalize newlines
             checked_content = []
             for i, content in enumerate(value):
-                if not isinstance(content, str):
+                fixer_query = ""
+                if isinstance(content, str) or isinstance(content, list):
+                    if isinstance(content, list):
+                        content, fixer_query = content
+
+                        if not isinstance(content, str) or not isinstance(
+                            fixer_query, str
+                        ):
+                            yield InterruptingError, {
+                                "message": (
+                                    "The '[content-to-exclude, fixer_query]' array"
+                                    f"  items '{pprint.pformat([content, fixer_query])}'"
+                                    " must be of type string"
+                                ),
+                                "definition": f".includeLines[{i}]",
+                            }
+                            return
+                else:
                     yield InterruptingError, {
                         "message": (
                             "The content to exclude"
                             f" '{pprint.pformat(content)}'"
-                            " must be of type string"
+                            " must be of type string or array"
                         ),
                         "definition": f".excludeContent[{i}]",
                         "file": fpath,
                     }
                     return
-                elif not content:
+
+                if not content:
                     yield InterruptingError, {
                         "message": "The content to exclude must not be empty",
                         "definition": f".excludeContent[{i}]",
@@ -276,12 +368,53 @@ class InclusionPlugin:
                     return
 
                 if content in fcontent:
+                    if fixer_query:
+                        fixable = True
+                        fixed = False
+                        if context.fix:
+                            try:
+                                compiled_fixer_query = (
+                                    compile_JMESPath_expression_or_error(
+                                        fixer_query,
+                                    )
+                                )
+                            except JMESPathError as exc:
+                                yield InterruptingError, {
+                                    "message": exc.message,
+                                    "definition": f".excludeContent[{i}]",
+                                }
+                                continue
+
+                            _, instance = tree.serialize_file(fpath)
+
+                            try:
+                                diff = fix_tree_serialized_file_by_jmespath(
+                                    compiled_fixer_query,
+                                    instance,
+                                    fpath,
+                                    tree,
+                                )
+                            except JMESPathError as exc:
+                                yield InterruptingError, {
+                                    "message": exc.message,
+                                    "definition": f".excludeContent[{i}]",
+                                }
+                                continue
+                            else:
+                                fixed = True
+                                if not diff:
+                                    continue
+                    else:
+                        fixed = False
+                        fixable = False
                     yield Error, {
                         "message": (
                             f"Found expected content to exclude '{content}'"
                         ),
                         "file": fpath,
                         "definition": f".excludeContent[{i}]",
+                        "fixed": fixed,
+                        "fixable": fixable,
                     }
                 else:
                     checked_content.append(content)
