+++ src/project_config/serializers/editorconfig.py
@@ -106,6 +106,23 @@ def loads(string: str) -> EditorConfigConfigType:
     return result
 
 
+def _pyobject_to_ini_str(obj: t.Any) -> str:
+    """Converts a Python object to a string.
+
+    Args:
+        obj (Any): Python object.
+
+    Returns:
+        str: String representation of the object.
+    """
+    if isinstance(obj, str):
+        if not obj:
+            return '""'
+    elif isinstance(obj, bool):
+        return "true" if obj else "false"
+    return str(obj)
+
+
 def dumps(obj: t.Any) -> str:
     """Converts a JSON object to a .editorconfig configuration file string.
 
@@ -117,10 +134,10 @@ def dumps(obj: t.Any) -> str:
     for key, value in obj.items():
         if key == "":
             for optname, optvalue in value.items():
-                result += f"{optname} = {optvalue}\n"
+                result += f"{optname} = {_pyobject_to_ini_str(optvalue)}\n"
         else:
             result += f"\n[{key}]\n"
             for optname, optvalue in value.items():
-                result += f"{optname} = {optvalue}\n"
+                result += f"{optname} = {_pyobject_to_ini_str(optvalue)}\n"
 
     return result
