+++ tests/test_unit/test_cli/test_commands/test_init.py
@@ -20,7 +20,10 @@ def test_init_default(tmp_path, chdir, capsys):
         style_file = tmp_path / "style.json5"
         assert style_file.exists()
 
-        assert run(["check"]) == 0
+        assert run(["fix"]) == 0
+        out, err = capsys.readouterr()
+        assert err == ""
+        assert out == ""
 
 
 def test_init_pyproject_toml(tmp_path, chdir, capsys):
@@ -43,7 +46,10 @@ def test_init_pyproject_toml(tmp_path, chdir, capsys):
         style_file = tmp_path / "style.json5"
         assert style_file.exists()
 
-        assert run(["check"]) == 0
+        assert run(["fix"]) == 0
+        out, err = capsys.readouterr()
+        assert err == ""
+        assert out == ""
 
 
 def test_init_custom_file(tmp_path, chdir, capsys):
@@ -63,7 +69,12 @@ def test_init_custom_file(tmp_path, chdir, capsys):
         style_file = tmp_path / "style.json5"
         assert style_file.exists()
 
-        assert run(["check", "-c", "custom-file.toml"]) == 0
+        exitcode = run(["fix", "--config", "custom-file.toml"])
+        out, err = capsys.readouterr()
+        msg = f"{out}\n---\n{err}"
+        assert err == "", msg
+        assert out == "", msg
+        assert exitcode == 0
 
 
 @pytest.mark.parametrize(
@@ -119,7 +130,10 @@ cache = "5 minutes"
         style_file = tmp_path / "style.json5"
         assert style_file.exists()
 
-        assert run(["check"]) == 0
+        assert run(["fix"]) == 0
+        out, err = capsys.readouterr()
+        assert err == ""
+        assert out == ""
 
 
 def test_toml_file_already_exists_and_section_found(tmp_path, chdir, capsys):
@@ -137,3 +151,86 @@ def test_toml_file_already_exists_and_section_found(tmp_path, chdir, capsys):
             "The configuration for project-config has already been"
             " initialized at pyproject.toml[tool.project-config]\n"
         )
+
+
+def test_project_config_ini_is_fixable(tmp_path, chdir, capsys):
+    with chdir(tmp_path):
+        project_config_file = tmp_path / ".project-config.toml"
+        style_file = tmp_path / "style.json5"
+
+        assert run(["init"]) == 0
+        out, err = capsys.readouterr()
+        msg = f"{out}\n---\n{err}"
+        assert err == "", msg
+        assert out == "Configuration initialized at .project-config.toml\n", msg
+
+        expected_project_config_content = (
+            'style = ["style.json5"]\ncache = "5 minutes"\n'
+        )
+        assert (
+            project_config_file.read_text() == expected_project_config_content
+        )
+        project_config_file.write_text("")
+        assert style_file.exists()
+
+        config_file = tmp_path / "custom-file.toml"
+        config_file.write_text('style = ["style.json5"]\n')
+
+        exitcode = run(["fix", "--config", "custom-file.toml", "--no-color"])
+        assert (
+            project_config_file.read_text() == expected_project_config_content
+        )
+        out, err = capsys.readouterr()
+        msg = f"{out}\n---\n{err}"
+        assert err.count("(FIXED)") == 3, msg
+        assert out == "", msg
+        assert exitcode == 1, msg
+
+        exitcode = run(["fix", "--config", "custom-file.toml", "--no-color"])
+        out, err = capsys.readouterr()
+        msg = f"{out}\n---\n{err}"
+        assert err == "", msg
+        assert out == "", msg
+        assert exitcode == 0, msg
+
+
+def test_pyproject_toml_ini_is_fixable(tmp_path, chdir, capsys):
+    with chdir(tmp_path):
+        pyproject_toml_file = tmp_path / "pyproject.toml"
+        style_file = tmp_path / "style.json5"
+
+        assert run(["init", "-c", "pyproject.toml"]) == 0
+        out, err = capsys.readouterr()
+        msg = f"{out}\n---\n{err}"
+        assert err == "", msg
+        assert (
+            out
+            == "Configuration initialized at pyproject.toml[tool.project-config]\n"
+        ), msg
+
+        expected_pyproject_toml_content = '[tool.project-config]\nstyle = ["style.json5"]\ncache = "5 minutes"\n'
+        assert (
+            pyproject_toml_file.read_text() == expected_pyproject_toml_content
+        )
+        pyproject_toml_file.write_text("")
+        assert style_file.exists()
+
+        config_file = tmp_path / "custom-file.toml"
+        config_file.write_text('style = ["style.json5"]\n')
+
+        exitcode = run(["fix", "--config", "custom-file.toml", "--no-color"])
+        assert (
+            pyproject_toml_file.read_text() == expected_pyproject_toml_content
+        )
+        out, err = capsys.readouterr()
+        msg = f"{out}\n---\n{err}"
+        assert err.count("(FIXED)") == 5, msg
+        assert out == "", msg
+        assert exitcode == 1, msg
+
+        exitcode = run(["fix", "--config", "custom-file.toml", "--no-color"])
+        out, err = capsys.readouterr()
+        msg = f"{out}\n---\n{err}"
+        assert err == "", msg
+        assert out == "", msg
+        assert exitcode == 0, msg
