+++ tests/test_unit/test_cli/test_commands/test_show_plugins.py
@@ -1,23 +1,33 @@
+import json
+
 from project_config.__main__ import run
 from project_config.plugins import Plugins
 
 
-def test_show_plugins(capsys):
-    assert run(["show", "plugins", "--nocolor"]) == 0
-    out, err = capsys.readouterr()
-    assert err == ""
+def test_show_plugins(tmp_path, chdir, capsys):
 
-    stdout_lines = out.splitlines()
+    with chdir(tmp_path):
+        project_config_file = tmp_path / ".project-config.toml"
+        project_config_file.write_text('style = "style.json"')
 
-    expected_plugins_names = Plugins().plugin_names
+        style_file = tmp_path / "style.json"
+        style_file.write_text(json.dumps({"rules": [{"files": "style.json"}]}))
 
-    outputted_plugin_names = []
-    for i, line in enumerate(stdout_lines):
-        if not line.startswith("  "):
-            plugin_name = line.rstrip(":")
-            assert plugin_name in expected_plugins_names
-            assert stdout_lines[i + 1].startswith("  - ")
-            outputted_plugin_names.append(plugin_name)
+        assert run(["show", "plugins", "--nocolor"]) == 0
+        out, err = capsys.readouterr()
+        assert err == ""
 
-    for expected_plugin_name in expected_plugins_names:
-        assert expected_plugin_name in outputted_plugin_names
+        stdout_lines = out.splitlines()
+
+        expected_plugins_names = Plugins().plugin_names
+
+        outputted_plugin_names = []
+        for i, line in enumerate(stdout_lines):
+            if not line.startswith("  "):
+                plugin_name = line.rstrip(":")
+                assert plugin_name in expected_plugins_names
+                assert stdout_lines[i + 1].startswith("  - ")
+                outputted_plugin_names.append(plugin_name)
+
+        for expected_plugin_name in expected_plugins_names:
+            assert expected_plugin_name in outputted_plugin_names
